Update builder to reflect changes to Arch build process.

+++ utils.py
@@ -30,7 +30,7 @@ import urllib.request, urllib.error, urllib.parse
 
 
 APP_NAME = 'archbuilder'
-BUILDER_USER = APP_NAME
+BUILDER_USER = 'nobody'
 ETC_LOCALE_GEN = '''
 en_US.UTF-8 UTF-8
 en_US ISO-8859-1
@@ -81,6 +81,9 @@ def UserExists(username):
 
 
 def CreateBuildUser(user=BUILDER_USER):
+  if user == BUILDER_USER:
+    logging.warning('Skipping create user.')
+    return
   if not UserExists(user):
     home_dir = '/home/%s' % user
     Run(['useradd', user, '-d', home_dir])
@@ -89,6 +92,9 @@ def CreateBuildUser(user=BUILDER_USER):
 
 
 def RemoveBuildUser():
+  if True:
+    logging.warning('Skipping delete user.')
+    return
   if UserExists(BUILDER_USER):
     Run(['userdel', '-r', BUILDER_USER])
 
@@ -272,6 +278,8 @@ def AurInstall(name=None, pkbuild_url=None):
   tarball = glob.glob(os.path.join(workspace_dir, '*.tar*'))
   tarball = tarball[0]
   Pacman(['-U', tarball], cwd=workspace_dir)
+  
+  RemoveBuildUser()
   return tarball
 
 
@@ -282,6 +290,13 @@ def Pacstrap(base_dir, params):
 def Pacman(params, cwd=None):
   Run(['pacman', '--noconfirm'] + params, cwd=cwd)
 
+
+def UpdatePacmanDatabase():
+  Pacman(['-Sy'])
+
+
+def UpdateAllPackages():
+  Pacman(['-Syyu'])
 
 def InstallPackages(package_list):
   Pacman(['-S'] + package_list)
@@ -304,6 +319,9 @@ class ImageMapper(object):
   def _LoadPartitionsIfNeeded(self):
     if not self._device_map:
       self.LoadPartitions()
+
+  def InstallLoopback(self):
+    SudoRun(['modprobe', 'loop'])
 
   def LoadPartitions(self):
     return_code, out, err = SudoRun(['kpartx', '-l', self._raw_disk], capture_output=True)
